Refactor API startup: modularize endpoints and seeding
Move user/economic endpoints to extension methods for clarity. Extract DB migration/admin seeding and connection string logic into separate extension classes. Clean up Program.cs and remove duplicate service registrations. Improves maintainability and code organization.
