+++ BetonBon.Client/Program.cs
@@ -36,10 +36,7 @@ namespace BetonBon.Client
             builder.Services.AddScoped<AuthenticationStateProvider, CustomAuthStateProvider>();
             builder.Services.AddScoped<LocalStorage>();
             builder.Services.AddTransient<AuthHeaderHandler>();
-
-            builder.Services.AddScoped<LocalStorage>();
             builder.Services.AddScoped<PopupService>();
-
 
             await builder.Build().RunAsync();
         }
